More fixes and content
Added some skills, education, and a new project.

Project pages are now a little smarter, showing the "back" button only if the referrer is the mainpage.

Changed monospace font for Roboto Mono, and fixed a couple other style quirks.

Code formatter now has HTML/JS/CSS, and is a little smarter with comments.

Thumbnails are back on mobile (very small)